update company search results
