+++ resources/js/components/organisms/CompanySearchResults.vue
@@ -1,33 +1,32 @@
 <script lang="ts" setup>
-    import { AisInfiniteHits, AisHighlight } from 'vue-instantsearch/vue3/es'
-    import CompanySearchResultCard from '../molecules/CompanySearchResultCard.vue';
-    import NoResults from '../atoms/NoResults.vue';
-    import Button from '../atoms/Button.vue'
+import { AisIndex, AisInfiniteHits, AisHighlight } from 'vue-instantsearch/vue3/es'
+import CompanySearchResultsItem from '../molecules/CompanySearchResultsItem.vue';
+import NoResults from '../atoms/NoResults.vue';
+import Button from '../atoms/Button.vue'
+const firstThreeItems = (items: any[]) => items.slice(0, 3)
 </script>
 
 <template>
-    <AisInfiniteHits>
-        <template #default="{
-            items,
-            refineNext,
-            isLastPage
-        }">
-            <ul>
-                <li v-for="company, key in items" :key="key">
-                    <CompanySearchResultCard :company-name=company.name :company-logo-url="`https://logo.clearbit.com/${company.url}`" :href="`/companies/${company.id}`">
-                        <template #companyName>
-                            <AisHighlight :hit="company" attribute="name" />
-                        </template>
-                        <template #associatedContactsNumber>
-                            {{ company.number_of_contacts }}
-                        </template>
-                    </CompanySearchResultCard>
-                </li>
-                <li class="text-center mt-4">
-                    <Button @click="refineNext" button-text="See more companies" :disabled="isLastPage"/>
-                </li>
-            </ul>
-        </template>
+  <AisIndex index-name="companies">
+    <AisInfiniteHits class="w-full">
+      <template #default="{
+        items,
+        refineNext,
+        isLastPage
+      }">
+        <ul class="grid grid-cols-3 gap-3">
+          <li v-for="company in firstThreeItems(items)" :key="company.id">
+            <CompanySearchResultsItem :company-name=company.name
+              :company-logo-url="`https://logo.clearbit.com/${company.url}`" :contact-count="company.number_of_contacts"
+              :href="`/companies/${company.id}`">
+              <template #name>
+                <AisHighlight :hit="company" attribute="name" />
+              </template>
+            </CompanySearchResultsItem>
+          </li>
+        </ul>
+      </template>
     </AisInfiniteHits>
-    <NoResults/>
+    <!-- <NoResults /> -->
+  </AisIndex>
 </template>
